Don't use the version argument after the step one.

+++ c2cgeoportal/scripts/c2ctool.py
@@ -36,6 +36,7 @@ import shutil
 import argparse
 import httplib2
 import yaml
+import pkg_resources
 from subprocess import check_call, CalledProcessError
 from argparse import ArgumentParser
 from alembic.config import Config
@@ -145,7 +146,7 @@ def _fill_arguments(command):
             "--step", type=int, help=argparse.SUPPRESS, default=0
         )
         parser.add_argument(
-            "version", metavar="VERSION", help="Upgrade to version"
+            "version", metavar="VERSION", nargs='?', help="Upgrade to version"
         )
     elif command == "deploy":
         parser.add_argument(
@@ -241,6 +242,12 @@ class C2cTool:
             self.print_step(0, intro="Fix it and run again the upgrade:")
             exit(1)
 
+        if self.options.version is None:
+            print("")
+            print(self.color_bar)
+            print("The VERSION environment variable is required for this upgrade step")
+            exit(1)
+
         if re.match(VERSION_RE, self.options.version) is not None:
             http = httplib2.Http()
             headers, _ = http.request(
@@ -288,6 +295,12 @@ class C2cTool:
 
     def step1(self):
         notes = []
+
+        if self.options.version is None:
+            print("")
+            print(self.color_bar)
+            print("The VERSION environment variable is required for this upgrade step")
+            exit(1)
 
         check_call(["git", "reset", "--hard"])
         check_call(["git", "clean", "--force", "-d"])
@@ -447,7 +460,9 @@ class C2cTool:
         self.print_step(4, intro="Then to commit your changes type:")
 
     def step4(self):
-        check_call(["git", "commit", "-m", "Upgrade to GeoMapFish %s" % self.options.version])
+        check_call(["git", "commit", "-m", "Upgrade to GeoMapFish {}".format(
+            pkg_resources.get_distribution("c2cgeoportal").version
+        )])
 
         print("")
         print(self.color_bar)
